backend/requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# LangChain
langchain>=0.0.300
langchain-google-genai>=0.0.2
langchain_chroma>=0.0.1

# Vector Store
chromadb>=0.4.0

# Embeddings
jina>=0.13.9
sentence-transformers>=2.2.2

# LLM
google-generativeai>=0.3.0

# Database
redis>=4.5.0

# News Processing
news-please>=1.5.0

# Async
httpx>=0.24.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
